# rootmulti/internal/maps: remove duplicated code, simplify and speed up KVPPair